requirements.txt
sentence-transformers
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
transformers 
optimum
llama-index-embeddings-huggingface-optimum

llama-index-core

llama-index-readers-file
llama-index-llms-ollama
llama-index-llms-groq

llama-index-readers-wikipedia
wikipedia
